cli: improve tsdump datadog upload
Previously, tsdump datadog upload was time consuming. This was due to
synchronous upload requests to datadog. Along with this, we were decoding raw
data to TimeSeriesData. Then again converting it to DatadogSeries. This change
addresse both challenges. We are now decoding timeseries data to DatadogSeries.
We have decoupled deocding & uploading data so that we will perform both
operations concurrently.

Epic: CC-27740
Release note: None
