+++ pkg/cli/tsdump_upload.go
@@ -1,0 +1,341 @@
+// Copyright 2020 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package cli
+
+import (
+	"bytes"
+	"compress/gzip"
+	"encoding/gob"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"sync"
+
+	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/server"
+	"github.com/cockroachdb/cockroach/pkg/ts"
+	"github.com/cockroachdb/cockroach/pkg/util/httputil"
+	"github.com/cockroachdb/cockroach/pkg/util/retry"
+	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+	"github.com/cockroachdb/errors"
+)
+
+const (
+	DatadogSeriesTypeUnknown = iota
+	DatadogSeriesTypeCounter
+	DatadogSeriesTypeRate
+	DatadogSeriesTypeGauge
+)
+
+var (
+	// each site in datadog has a different host name. ddSiteToHostMap
+	// holds the mapping of site name to the host name.
+	ddSiteToHostMap = map[string]string{
+		"us1":     "api.datadoghq.com",
+		"us3":     "api.us3.datadoghq.com",
+		"us5":     "api.us5.datadoghq.com",
+		"eu1":     "api.datadoghq.eu",
+		"ap1":     "api.ap1.datadoghq.com",
+		"us1-fed": "api.ddog-gov.com",
+	}
+
+	targetURLFormat = "https://%s/api/v2/series"
+)
+
+// DatadogPoint is a single metric point in Datadog format
+type DatadogPoint struct {
+	// Timestamp must be in seconds since Unix epoch.
+	Timestamp int64   `json:"timestamp"`
+	Value     float64 `json:"value"`
+}
+
+// DatadogSeries contains a JSON encoding of a single series object
+// that can be send to Datadog.
+type DatadogSeries struct {
+	Metric    string         `json:"metric"`
+	Type      int            `json:"type"`
+	Points    []DatadogPoint `json:"points"`
+	Resources []struct {
+		Name string `json:"name"`
+		Type string `json:"type"`
+	} `json:"resources"`
+	// In order to encode arbitrary key-value pairs, use a `:` delimited
+	// tag string like `cluster:dedicated`.
+	Tags []string `json:"tags"`
+}
+
+// DatadogSubmitMetrics is the top level JSON object that must be sent to Datadog.
+// See: https://docs.datadoghq.com/api/latest/metrics/#submit-metrics
+type DatadogSubmitMetrics struct {
+	Series []DatadogSeries `json:"series"`
+}
+
+type DatadogResp struct {
+	Errors []string `json:"errors"`
+}
+
+var newTsdumpUploadID = func() string {
+	clusterTagValue := ""
+	if debugTimeSeriesDumpOpts.clusterLabel != "" {
+		clusterTagValue = debugTimeSeriesDumpOpts.clusterLabel
+	} else if serverCfg.ClusterName != "" {
+		clusterTagValue = serverCfg.ClusterName
+	} else {
+		clusterTagValue = fmt.Sprintf("cluster-debug-%d", timeutil.Now().Unix())
+	}
+	return newUploadID(clusterTagValue)
+}
+
+// datadogWriter can convert our metrics to Datadog format and send
+// them via HTTP to the public DD endpoint, assuming an API key is set
+// in the CLI flags.
+type datadogWriter struct {
+	sync.Once
+	targetURL string
+	uploadID  string
+	init      bool
+	apiKey    string
+	// namePrefix sets the string to prepend to all metric names. The
+	// names are kept with `.` delimiters.
+	namePrefix string
+	doRequest  func(req *http.Request) error
+	threshold  int
+}
+
+func makeDatadogWriter(
+	targetURL string,
+	init bool,
+	apiKey string,
+	threshold int,
+	doRequest func(req *http.Request) error,
+) *datadogWriter {
+	return &datadogWriter{
+		targetURL:  targetURL,
+		uploadID:   newTsdumpUploadID(),
+		init:       init,
+		apiKey:     apiKey,
+		namePrefix: "crdb.tsdump.", // Default pre-set prefix to distinguish these uploads.
+		doRequest:  doRequest,
+		threshold:  threshold,
+	}
+}
+
+func doDDRequest(req *http.Request) error {
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+	respBytes, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+	ddResp := DatadogResp{}
+	err = json.Unmarshal(respBytes, &ddResp)
+	if err != nil {
+		return err
+	}
+	if len(ddResp.Errors) > 0 {
+		return errors.Newf("tsdump: error response from datadog: %v", ddResp.Errors)
+	}
+	if resp.StatusCode > 299 {
+		return errors.Newf("tsdump: bad response status code: %+v", resp)
+	}
+	return nil
+}
+
+func dump(kv *roachpb.KeyValue) (*DatadogSeries, error) {
+	name, source, _, _, err := ts.DecodeDataKey(kv.Key)
+	if err != nil {
+		return nil, err
+	}
+	var idata roachpb.InternalTimeSeriesData
+	if err := kv.Value.GetProto(&idata); err != nil {
+		return nil, err
+	}
+
+	series := &DatadogSeries{
+		Metric: name,
+		Tags:   []string{},
+		Type:   DatadogSeriesTypeUnknown,
+		Points: make([]DatadogPoint, idata.SampleCount()),
+	}
+
+	sl := reCrStoreNode.FindStringSubmatch(name)
+	if len(sl) != 0 {
+		storeNodeKey := sl[1]
+		if storeNodeKey == "node" {
+			storeNodeKey += "_id"
+		}
+		series.Tags = append(series.Tags, fmt.Sprintf("%s:%s", storeNodeKey, source))
+		series.Metric = sl[2]
+	} else {
+		series.Tags = append(series.Tags, "node_id:0")
+	}
+
+	for i := 0; i < idata.SampleCount(); i++ {
+		if idata.IsColumnar() {
+			series.Points[i].Timestamp = idata.TimestampForOffset(idata.Offset[i]) / 1_000_000_000
+			series.Points[i].Value = idata.Last[i]
+		} else {
+			series.Points[i].Timestamp = idata.TimestampForOffset(idata.Samples[i].Offset) / 1_000_000_000
+			series.Points[i].Value = idata.Samples[i].Sum
+		}
+
+	}
+	return series, nil
+}
+
+func (d *datadogWriter) emitDataDogMetrics(data []DatadogSeries) error {
+	var tags []string
+	// Hardcoded values
+	tags = append(tags, "cluster_type:SELF_HOSTED")
+	tags = append(tags, "job:cockroachdb")
+	tags = append(tags, "region:local")
+
+	if debugTimeSeriesDumpOpts.clusterLabel != "" {
+		tags = append(tags, makeDDTag("cluster_label", debugTimeSeriesDumpOpts.clusterLabel))
+	}
+
+	tags = append(tags, makeDDTag(uploadIDTag, d.uploadID))
+
+	d.Do(func() {
+		fmt.Println("Upload ID:", d.uploadID)
+	})
+
+	for i := 0; i < len(data); i++ {
+		data[i].Tags = append(data[i].Tags, tags...)
+		data[i].Metric = d.namePrefix + data[i].Metric
+	}
+
+	// When running in init mode, we insert zeros with the current
+	// timestamp in order to populate Datadog's metrics list. Then the
+	// user can enable these metrics for historic ingest and load the
+	// full dataset. This should only be necessary once globally.
+	if d.init {
+		for i := 0; i < len(data); i++ {
+			data[i].Points = []DatadogPoint{{
+				Value:     0,
+				Timestamp: timeutil.Now().Unix(),
+			}}
+		}
+	}
+
+	fmt.Printf(
+		"tsdump datadog upload: sending payload containing %d series including %s\n",
+		len(data),
+		data[0].Metric,
+	)
+
+	return d.flush(data)
+}
+
+func (d *datadogWriter) flush(data []DatadogSeries) error {
+	var buf bytes.Buffer
+	err := json.NewEncoder(&buf).Encode(&DatadogSubmitMetrics{Series: data})
+	if err != nil {
+		return err
+	}
+	var zipBuf bytes.Buffer
+	g := gzip.NewWriter(&zipBuf)
+	_, err = io.Copy(g, &buf)
+	if err != nil {
+		return err
+	}
+	err = g.Close()
+	if err != nil {
+		return err
+	}
+
+	retryOpts := base.DefaultRetryOptions()
+	retryOpts.MaxRetries = 5
+	for retry := retry.Start(retryOpts); retry.Next(); {
+		req, err := http.NewRequest("POST", d.targetURL, &zipBuf)
+		if err != nil {
+			return err
+		}
+		req.Header.Set("DD-API-KEY", d.apiKey)
+		req.Header.Set(server.ContentTypeHeader, "application/json")
+		req.Header.Set(httputil.ContentEncodingHeader, "gzip")
+
+		err = d.doRequest(req)
+		if err == nil {
+			return nil
+		}
+	}
+	return err
+}
+
+func (d *datadogWriter) upload(fileName string) error {
+	f, err := os.Open(fileName)
+	if err != nil {
+		return err
+	}
+
+	dec := gob.NewDecoder(f)
+	gob.Register(&roachpb.KeyValue{})
+	decodeOne := func() ([]DatadogSeries, error) {
+		var ddSeries []DatadogSeries
+
+		for i := 0; i < d.threshold; i++ {
+			var v roachpb.KeyValue
+			err := dec.Decode(&v)
+			if err != nil {
+				return ddSeries, err
+			}
+
+			datadogSeries, err := dump(&v)
+			if err != nil {
+				return nil, err
+			}
+			ddSeries = append(ddSeries, *datadogSeries)
+		}
+
+		return ddSeries, nil
+	}
+
+	var wg sync.WaitGroup
+	ch := make(chan []DatadogSeries, 4000)
+
+	for i := 0; i < 1000; i++ {
+		go func() {
+			for data := range ch {
+				err := d.emitDataDogMetrics(data)
+				if err != nil {
+					fmt.Printf("retries exhausted for datadog upload containing series %s with error %v", data[0].Metric, err)
+					wg.Done()
+					return
+				}
+				wg.Done()
+			}
+		}()
+	}
+
+	for {
+		data, err := decodeOne()
+		if err == io.EOF {
+			if len(data) != 0 {
+				wg.Add(1)
+				ch <- data
+			}
+			break
+		}
+		wg.Add(1)
+		ch <- data
+	}
+
+	wg.Wait()
+	close(ch)
+	return nil
+}
